moved getEmailWithOptionalName to storage module
